refactor: extract lesson editor helpers
